SbAudio: remove vim modelines
These were not in every file and I guess I had bumped into that
because when I started editing code I immediately ran into tabs vs spaces
issues.

The decision is thus to either add more of these directives,
or to move to a more modern encoding for indentation. I've picked the latter.