SbAudio: pull in libsoundio
The current plan is for libsoundio to be the only backend.

I also want its source code in the same repo instead
of depending on external sources.

At the moment I'm also going to be opinionated in the
way libsoundio is integrated, but this can easily
be reverted later to using standard libsoundio,
though hosted in this codebase for ease of use.

SDL2 and Windows backends will be removed once a libsoundio
backend has been written.