SbAudio: remove msvc windows GH Actions runner for now
libsoundio is using <stdatomic.h> and I'm having trouble getting this to build on Windows.
Windows MSVC is very far behind and they just added experimental support for that header
in late 2022.

Instead I've added a Windows Clang runner.

Temporary solution

Also removes mingw runner as it's not a target platform for this project.